// features/chinese_chess_hidden/models/ChineseChessName.ts
export enum ChessNameBlack {
  King = '將',
  Guard = '士',
  Minister = '象',
  Chariot = '車',
  Cannon = '砲',
  Horse = '馬',
  Soldier = '卒',
}

export enum ChessNameRed {
  King = '帥',
  Guard = '仕',
  Minister = '相',
  Chariot = '俥',
  Cannon = '炮',
  Horse = '傌',
  Soldier = '兵',
}